Deploy a placeholder zip when creating lambda resources
We should not be deploying via terraform. It's possible that unreviewed or out of date code may be deployed unintentionally this way.
We have a working CI/CD pipeline to deploy lambda code so disable terraform deployment of lambda code in favour of deploying a placeholder zip.
This will need to be followed by a CI/CD run but we have a manual trigger in GH Actions to allow us to do this.