Add a manual workflow to allow for oracle client layer creation and publishing
This workflow downloads the appropriate Oracle client archive and builds a lambda layer we can use in thick client mode with Oracle.
This is a manual workflow because we really only need to do this once in the environment setup. We could arguably do it in Terraform when creating resources.